requirements.txt
openai>=1.40.0
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pytest>=8.0.0
